api/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
requests
PyJWT
pydantic[email]
